Moved development tasks to separate directory.
