--- dev/tasks/dev.js
@@ -1,52 +1,0 @@
-/**
- * @license Copyright (c) 2003-2016, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md.
- */
-
-'use strict';
-
-const initTask = require( './utils/dev-init' );
-const pluginCreateTask = require( './utils/dev-plugin-create' );
-const pluginUpdateTask = require( './utils/dev-update' );
-const pluginStatusTask = require( './utils/dev-status' );
-const installTask = require( './utils/dev-install' );
-const relinkTask = require( './utils/dev-relink' );
-const boilerplateUpdateTask = require( './utils/dev-boilerplate-update' );
-const ckeditor5Path = process.cwd();
-
-module.exports = ( grunt ) => {
-	const packageJSON = grunt.config.data.pkg;
-	const workspaceRoot = grunt.config.data.workspaceRoot;
-
-	grunt.registerTask( 'dev-init', function() {
-		initTask( installTask, ckeditor5Path, packageJSON, workspaceRoot, grunt.log.writeln );
-	} );
-
-	grunt.registerTask( 'dev-plugin-create', function() {
-		const done = this.async();
-		pluginCreateTask( ckeditor5Path, workspaceRoot, grunt.log.writeln )
-			.then( done )
-			.catch( ( error )  => done( error ) );
-	} );
-
-	grunt.registerTask( 'dev-update', function() {
-		pluginUpdateTask( ckeditor5Path, packageJSON, workspaceRoot, grunt.log.writeln, grunt.option( 'npm-update' ) );
-	} );
-
-	grunt.registerTask( 'dev-status', function() {
-		pluginStatusTask( ckeditor5Path, packageJSON, workspaceRoot, grunt.log.writeln, grunt.log.error );
-	} );
-
-	grunt.registerTask( 'dev-boilerplate-update', function() {
-		boilerplateUpdateTask( ckeditor5Path, packageJSON, workspaceRoot, grunt.log.writeln, grunt.log.error );
-	} );
-
-	grunt.registerTask( 'dev-relink', function() {
-		relinkTask( ckeditor5Path, packageJSON, workspaceRoot, grunt.log.writeln, grunt.log.error );
-	} );
-
-	grunt.registerTask( 'dev-install', function( ) {
-		installTask( ckeditor5Path, workspaceRoot, grunt.option( 'plugin' ), grunt.log.writeln );
-	} );
-};
-
