+++ dev/tasks/gulp/dev/tasks/dev-boilerplate-update.js
@@ -5,13 +5,14 @@
 
 'use strict';
 
-const tools = require( './tools' );
+const tools = require( '../utils/tools' );
+const git = require( '../utils/git' );
 const path = require( 'path' );
 
 /**
  * 1. Get CKEditor5 dependencies from package.json file.
  * 2. Scan workspace for repositories that match dependencies from package.json file.
- * 3. Link repositories to node_modules in CKEditor5 repository.
+ * 3. Fetch and merge boilerplate remote.
  *
  * @param {String} ckeditor5Path Path to main CKEditor5 repository.
  * @param {Object} packageJSON Parsed package.json file from CKEditor5 repository.
@@ -31,13 +32,12 @@ module.exports = ( ckeditor5Path, packageJSON, workspaceRoot, writeln, writeErro
 		if ( directories.length ) {
 			for ( let dependency in dependencies ) {
 				const repositoryAbsolutePath = path.join( workspaceAbsolutePath, dependency );
-				const repositoryURL = dependencies[ dependency ];
 
-				// Check if repository's directory exists.
+				// Check if repository's directory already exists.
 				if ( directories.indexOf( dependency ) > -1 ) {
 					try {
-						writeln( `Linking ${ repositoryURL }...` );
-						tools.linkDirectories( repositoryAbsolutePath, path.join( ckeditor5Path, 'node_modules', dependency ) );
+						writeln( `Updating boilerplate in ${ dependency }...` );
+						git.updateBoilerplate( repositoryAbsolutePath );
 					} catch ( error ) {
 						writeError( error );
 					}
